Time: 189 ms (70.22%), Space: 125.7 MB (29.12%) - LeetHub

+++ 295-find-median-from-data-stream/295-find-median-from-data-stream.java
@@ -24,7 +24,7 @@ class MedianFinder {
     public double findMedian() {
          if(even)
         {
-           return (minHeap.peek() + maxHeap.peek())/2.0;
+           return (double)(minHeap.peek() + maxHeap.peek())/2;
         }
         else
         {
